EShopProject/EShopWebParent/EShopFrontEnd/src/test/java/com/eshop/product/ProductRepositoryTests.java
package com.eshop.product;

import com.eshop.common.entity.Product;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

@DataJpaTest
@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
public class ProductRepositoryTests {

    @Autowired
    private ProductRepository productRepository;

    @Test
    public void testFindByAlias() {
        String alias = "something";
        Product product = productRepository.findByAlias(alias);

        assertThat(product).isNotNull();
    }
}
